requirements.txt
pandas
numpy
python-dotenv>=0.19.0
streamlit
scikit-learn